requirements.txt
black==25.1.0
blinker==1.9.0
cfgv==3.4.0
click==8.1.8
distlib==0.3.9
filelock==3.17.0
Flask==3.1.0
identify==2.6.7
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pre_commit==4.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
PyYAML==6.0.2
virtualenv==20.29.2
Werkzeug==3.1.3